FIX (retention): Fix GFS retention while hourly backups prevent daily from cleanup
